Adds api utility using axios, renames store folder to Store
