+++ package.json
@@ -31,6 +31,7 @@
     }
   },
   "devDependencies": {
+    "axios": "^0.18.0",
     "babel-core": "^6.26.3",
     "babel-eslint": "^ 9.0.0",
     "babel-loader": "^7.1.4",
